# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Merchant and OWNER created atomically"
          }
        },
        "summary": "Register merchant and its primary OWNER",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Login and start a rotating server-side session",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfileCompatibility",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/devices": {
      "post": {
        "operationId": "DevicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendaftarkan device baru (Owner)",
        "tags": [
          "Devices"
        ]
      },
      "get": {
        "operationId": "DevicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan daftar semua kasir (Owner)",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/devices/pair": {
      "post": {
        "operationId": "DevicesController_pairDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melakukan pairing menggunakan 6-digit kode (Public)",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/devices/{id}": {
      "delete": {
        "operationId": "DevicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mencabut akses / menghapus device (Owner)",
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/api/v1/payments": {
      "get": {
        "operationId": "PaymentsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/{id}/reconciliations": {
      "post": {
        "operationId": "PaymentsController_open",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenPaymentReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payment-reconciliations": {
      "get": {
        "operationId": "PaymentsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payment-reconciliations/{id}/resolve": {
      "post": {
        "operationId": "PaymentsController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolvePaymentReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "K-POS API",
    "description": "Canonical API for the K-POS multi-app offline-first platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "email": {
            "type": "string",
            "example": "budi@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8
          },
          "merchant_name": {
            "type": "string",
            "example": "Toko Kopi Budi",
            "description": "Nama toko/merchant"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Jakarta",
            "default": "Asia/Jakarta"
          }
        },
        "required": [
          "full_name",
          "email",
          "password",
          "merchant_name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "budi@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "device_id": {
            "type": "string",
            "description": "Required for OPERATOR login on a paired counter"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateDeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tablet Kasir Utama"
          }
        },
        "required": [
          "name"
        ]
      },
      "PairDeviceDto": {
        "type": "object",
        "properties": {
          "pairing_code": {
            "type": "string",
            "example": "123456"
          },
          "hardware_id": {
            "type": "string",
            "example": "SIMULASI-IPAD-1234"
          }
        },
        "required": [
          "pairing_code",
          "hardware_id"
        ]
      },
...
      "OpenPaymentReconciliationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5
          },
          "evidence_note": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "ResolvePaymentReconciliationDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "VALID",
              "INVALID"
            ]
          },
          "resolution_note": {
            "type": "string",
            "minLength": 5
          },
          "inventory_returned": {
            "type": "boolean",
            "description": "Only relevant for INVALID resolution"
          }
        },
        "required": [
          "action",
          "resolution_note"
        ]
      },
...
    }
  }
}